refactor(auth): always compile integration tests
Split the code such that the bulk of the tests are always compiled, even
if they do not get to run. That makes it easier to detect build problems
before sending a PR for review.
