+++ src/auth/integration-tests/Cargo.toml
@@ -23,13 +23,13 @@ publish           = false
 run-integration-tests = []
 
 [dependencies]
-
-[dev-dependencies]
 auth          = { path = "../../../src/auth", package = "google-cloud-auth" }
 gax           = { path = "../../../src/gax", package = "google-cloud-gax" }
 language      = { path = "../../../src/generated/cloud/language/v2", package = "google-cloud-language-v2" }
 scoped-env    = "2"
 secretmanager = { path = "../../../src/generated/cloud/secretmanager/v1", package = "google-cloud-secretmanager-v1" }
-serial_test   = "3"
 tempfile      = "3"
+
+[dev-dependencies]
+serial_test   = "3"
 tokio         = { version = "1", features = ["full", "macros"] }
